Use generic type instead of `Function` for cost function in some solvers
Make the method signatures of quasi_Newton, cyclic_proximal_point,a
and ChambollePock consistent with other solver functions such as for
example gradient_descent.
A generic type instead of `Function` allows the methods to be called
with a function like object. This pattern is for example demonstrated
in the geodesic regression tutorial.